finished adding authentication to all routes that require it. finished integrating relationship between users and tasks

+++ src/routers/task.js
@@ -18,10 +18,11 @@ router.post('/tasks', auth, async (req, res) => {
 })
 
 //Find all tasks
-router.get('/tasks', async (req, res) => {
+router.get('/tasks', auth, async (req, res) => {
     try {
-        const tasks = await Task.find({})
-        res.send(tasks)
+        await req.user.populate('tasks').execPopulate()
+
+        res.send(req.user.tasks)
     } catch (e) {
         res.status(500).send()
     }
@@ -42,11 +43,9 @@ router.get('/tasks/:id', auth, async (req, res) => {
         res.status(500).send()
     }
 })
-// left off on 2:50 lesson 115
-
 
 //Update task
-router.patch('/tasks/:id', async (req, res) => {
+router.patch('/tasks/:id', auth, async (req, res) => {
     const updates = Object.keys(req.body)
     const allowedUpdates = ['description', 'completed']
     const isValidOperation = updates.every((update) => allowedUpdates.includes(update))
@@ -56,14 +55,14 @@ router.patch('/tasks/:id', async (req, res) => {
     }
 
     try {
-        const task = await Task.findById(req.params.id)
-        updates.forEach((update) => task[update] = req.body[update])
-        await task.save()
-        // const task = await Task.findByIdAndUpdate(req.params.id, req.body, { new: true, runValidators: true })
+        const task = await Task.findOne({ _id: req.params.id, owner: req.user._id})
 
         if (!task) {
             res.status(404).send()
         }
+
+        updates.forEach((update) => task[update] = req.body[update])
+        await task.save()
         res.send(task)
     } catch (e) {
         res.status(400).send()
@@ -71,9 +70,9 @@ router.patch('/tasks/:id', async (req, res) => {
 })
 
 //Delete task
-router.delete('/tasks/:id', async (req, res) => {
+router.delete('/tasks/:id', auth, async (req, res) => {
     try {
-        const task = await Task.findByIdAndDelete(req.params.id)
+        const task = await Task.findByIdAndDelete({  _id: req.params.id, owner: req.user._id })
 
         if (!task) {
             res.status(404).send()
